Two versions of CoffeeScript. public/private functions.
The reason this commit is so massive is that ActionCoffee (better name
pending) is unable to compile itself, since it's written in
CoffeeScript, not ActionCoffee (better name pending). So it requires a
pristine version of CoffeeScript to compile itself.

Hence also Cakefile and Cakefile-pristine. Admittedly, everything is a
bit of a mess right now.